spark/requirements.txt
kafka-python==2.0.2
pymongo==4.6.1
pyspark==3.5.0